bring it up

+++ crates/arch/src/memory/page.rs
@@ -3,7 +3,7 @@ use core::ops::Not;
 // pub use super::dtype;
 
 mod dtype {
-    pub type Addr = u64;
+    pub type Addr = usize;
 }
 
 pub const SIZE: dtype::Addr = 0x1000;
